% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{addListing}
\alias{addListing}
\title{Add to the datastore contents listing}
\usage{
addListing(DatastoreListing_ls, DataListing_ls)
}
\arguments{
\item{DatastoreListing_ls}{a list having the elements described.}

\item{DataListing_ls}{a list having four elements that correspond to the
elements of the DatastoreListing_ls.}
}
\value{
a list having the same structure as DatastoreListing_ls with the
data in DataListing_ls added to it.
}
\description{
\code{addListing} adds the information for a new datastore component
(i.e. table collection, table, dataset) to the list which keeps track of
datastore contents.
}
\details{
A list (DatastoreListing_ls) stored in the 'DatastoreListing.Rda' file keeps
track of the contents of the datastore. This list has 4 named components as
follows:
1) group - an ordered vector of the names of directory paths where the
respective components are located. The directory paths use the forward slash
(/) to separate directories. The root directory of the datastore is
represented by a leading forward slash.
2) name - a vector of names of the datastore component corresponding the
group names.
3) groupname - a vector of the complete path names to each datastore
component in the datastore. This is a concatenation of the 'group' and 'name'
parts.
4) attributes - list of the attributes of each datastore component where
each element is a list.
This listing is updated whenever a new component is added to the datastore.
The information for that component is concatenated to end of each of the
elements of the datastore listing.
}
