% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{addWarningMsg}
\alias{addWarningMsg}
\title{Add a warning message to the results list}
\usage{
addWarningMsg(ResultsListName, WarnMsg)
}
\arguments{
\item{ResultsListName}{the name of the results list given as a character
string}

\item{WarnMsg}{a character string that contains the warning message}
}
\value{
None. The function modifies the results list by adding a warning
message to the Warnings component of the results list. It creates the
Warnings component if it does not already exist.
}
\description{
\code{addWarningMsg} a visioneval framework module developer function that
adds an warning message to the Warnings component of the module results list
that is passed back to the framework.
}
\details{
This function is a convenience function for module developers for passing
warning messages back to the framework. The preferred method for handling
warnings in module execution is for the module to handle the warning by
passing one or more warning messages back to the framework. The framework
will then write warning messages to the log and stop execution. Warning
messages are stored in a component of the returned list called Warnings. This
component is a string vector where each element is an warning message. The
addWarningMsg will create the Warning component if it does not already exist
and will add a warning message to the vector.
}
