% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{applyBinomialModel}
\alias{applyBinomialModel}
\title{Applies an estimated binomial model to a set of input values.}
\usage{
applyBinomialModel(Model_ls, Data_df, TargetProp = NULL,
  CheckTargetSearchRange = FALSE, ApplyRandom = TRUE)
}
\arguments{
\item{Model_ls}{a list which contains the following components:
'Type' which has a value of 'binomial';
'Formula' a string representation of the model equation;
'Choices' a two-element vector listing the choice set. The first element is
the choice that the binary logit model equation predicts the odds of;
'PrepFun' a function which prepares the input data frame for the model
application. If no preparation, this element of the list should not be
present or should be set equal to NULL;
'SearchRange' a two-element numeric vector which specifies the acceptable
search range to use when determining the factor for adjusting the model
constant.
'RepeatVar' a string which identifies the name of a field to use for
repeated draws of the model. This is used in the case where for example the
input data is households and the output is vehicles and the repeat variable
is the number of vehicles in the household.
'ApplyRandom' a logical identifying whether the results will be affected by
random draws (i.e. if a random number in range 0 - 1 is less than the
computed probability) or if a probability cutoff is used (i.e. if the
computed probability is greater then 0.5). This is an optional component. If
it isn't present, the function runs with ApplyRandom = TRUE.}

\item{Data_df}{a data frame containing the data required for applying the
model.}

\item{TargetProp}{a number identifying a target proportion for the default
choice to be achieved for the input data or NULL if there is no target
proportion to be achieved.}

\item{CheckTargetSearchRange}{a logical identifying whether the function
is to only check whether the specified 'SearchRange' for the model will
produce acceptable values (i.e. no NA or NaN values). If FALSE (the default),
the function will run the model and will not check the target search range.}

\item{ApplyRandom}{a logical identifying whether the outcome will be
be affected by random draws (i.e. if a random number in range 0 - 1 is less
than the computed probability) or if a probability cutoff is used (i.e. if
the computed probability is greater than 0.5)}
}
\value{
a vector of choice values for each record of the input data frame if
the model is being run, or if the function is run to only check the target
search range, a two-element vector identifying if the search range produces
NA or NaN values.
}
\description{
\code{applyBinomialModel} a visioneval framework module developer function
that applies an estimated binomial model to a set of input data.
}
\details{
The function calculates the result of applying a binomial logit model to a
set of input data. If a target proportion (TargetProp) is specified, the
function calls the 'binarySearch' function to calculate an adjustment to
the constant of the model equation so that the population proportion matches
the target proportion. The function will also test whether the target search
range specified for the model will produce acceptable values.
}
