% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{applyLinearModel}
\alias{applyLinearModel}
\title{Applies an estimated linear model to a set of input values.}
\usage{
applyLinearModel(Model_ls, Data_df, TargetMean = NULL,
  CheckTargetSearchRange = FALSE)
}
\arguments{
\item{Model_ls}{a list which contains the following components:
'Type' which has a value of 'linear';
'Formula' a string representation of the model equation;
'PrepFun' a function which prepares the input data frame for the model
application. If no preparation, this element of the list should not be
present or should be set equal to NULL;
'SearchRange' a two-element numeric vector which specifies the acceptable
search range to use when determining the dispersion factor.
'OutFun' a function that is applied to transform the results of applying the
linear model. For example to untransform a power-transformed variable. If
no transformation is necessary, this element of the list should not be
present or should be set equal to NULL.}

\item{Data_df}{a data frame containing the data required for applying the
model.}

\item{TargetMean}{a number identifying a target mean value to be achieved  or
NULL if there is no target.}

\item{CheckTargetSearchRange}{a logical identifying whether the function
is to only check whether the specified 'SearchRange' for the model will
produce acceptable values (i.e. no NA or NaN values). If FALSE (the default),
the function will run the model and will not check the target search range.}
}
\value{
a vector of numeric values for each record of the input data frame if
the model is being run, or if the function is run to only check the target
search range, a summary of predicted values when the model is run with
dispersion set at the high value of the search range.
}
\description{
\code{applyLinearModel} applies an estimated linear model to a set of input
data.
}
\details{
The function calculates the result of applying a linear regression model to a
set of input data. If a target mean value (TargetMean) is specified, the
function calculates a standard deviation of a sampling distribution which
is applied to linear model results. For each value returned by the linear
model, a sample is drawn from a normal distribution where the mean value of
the distribution is the linear model result and the standard deviation of the
distibution is calculated by the binary search to match the population mean
value to the target mean value. This process is meant to be applied to linear
model where the dependent variable is power transformed. Applying the
sampling distribution to the linear model results increases the dispersion
of results to match the observed dispersion and also matches the mean values
of the untransformed results. This also enables the model to be applied to
situations where the mean value is different than the observed mean value.
}
