% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{assignDatastoreFunctions}
\alias{assignDatastoreFunctions}
\title{Assign datastore interaction functions}
\usage{
assignDatastoreFunctions(DstoreType)
}
\arguments{
\item{DstoreType}{A string identifying the datastore type.}
}
\value{
None. The function assigns datastore interactions functions to the
first position of the search path.
}
\description{
\code{assignDatastoreFunctions} a visioneval framework control function that
assigns the values of the functions for interacting with the datastore to the
functions for the declared datastore type.
}
\details{
The visioneval framework can work with different types of datastores. For
example a datastore which stores datasets in an HDF5 file or a datastore
which stores datasets as RData files in a directory hierarchy. This function
reads the 'DatastoreType' parameter from the model state file and then
assigns the common datastore interaction functions the values of the
functions for the declared datastore type.
}
