% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{assignLoad}
\alias{assignLoad}
\title{Return a data file anonomously}
\usage{
assignLoad(Filename)
}
\arguments{
\item{Filename}{a string identifying the full path to the data file to be
loaded.}
}
\value{
the contents of the file identified by the Filename
}
\description{
\code{assignLoad} loads a data file and returns anonomously
}
\details{
This function loads a dataset and returns it anomously so that it can be
assigned to a name.
}
