% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{binarySearch}
\alias{binarySearch}
\title{Binary search function to find a parameter which achieves a target value.}
\usage{
binarySearch(Function, SearchRange_, ..., Target = 0, MaxIter = 100,
  Tolerance = 1e-04)
}
\arguments{
\item{Function}{a function which returns a value which is compared to the
'Target' argument. The function must take as its first argument a value which
from the 'SearchRange_'. It must return a value that may be compared to the
'Target' value.}

\item{SearchRange_}{a two element numeric vector which has the lowest and
highest values of the parameter range within which the search will be carried
out.}

\item{...}{one or more optional arguments for the 'Function'.}

\item{Target}{a numeric value that is compared with the return value of the
'Function'.}

\item{MaxIter}{an integer specifying the maximum number of iterations
to all the search to attempt.}

\item{Tolerance}{a numeric value specifying the proportional difference
between the 'Target' and the return value of the 'Function' to determine
when the search is complete.}
}
\value{
the value in the 'SearchRange_' for the function parameter which
matches the target value.
}
\description{
\code{binarySearch} uses a binary search algorithm to find the value of a
function parameter for which the function achieves a target value.
}
\details{
A binary search algorithm is used by several modules to calibrate the
intercept of a binary logit model to match a specified proportion or to
calibrate a dispersion parameter for a linear model to match a mean value.
This function implements a binary search algorithm in a consistent manner to
be used in all modules that need it. It is written to work with stochastic
models which by their nature don't produce the same outputs given the same
inputs and so will not converge reliably. To deal with the stochasticity,
this function uses a successive averaging  approach to smooth out the effect
of stochastic variation on reliable convergence. Rather than use the results
of a single search iteration to determine the next value range to use in the
search, a weighted average of previous values is used with the more recent
values being weighted more heavily.
}
