% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkDataConsistency}
\alias{checkDataConsistency}
\title{Check data consistency with specification}
\usage{
checkDataConsistency(DatasetName, Data_, DstoreAttr_)
}
\arguments{
\item{DatasetName}{A string identifying the dataset that is being checked.}

\item{Data_}{A vector of values that may be of type integer, double,
character, or logical.}

\item{DstoreAttr_}{A named list where the components are the attributes of a
dataset.}
}
\value{
A list containing two components, Errors and Warnings. If no
  inconsistencies are found, both components will have zero-length character
  vectors. If there are one or more inconsistencies, then these components
  will hold vectors of error and warning messages. Mismatch between UNITS
  will produce a warning message. All other inconsistencies will produce
  error messages.
}
\description{
\code{checkDataConsistency} a visioneval framework control function that
checks whether data to be written to a dataset is consistent with the dataset
attributes.
}
\details{
This function compares characteristics of data to be written to a dataset to
the dataset attributes to determine whether they are consistent.
}
