% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{checkGeography}
\alias{checkGeography}
\title{Check geographic specifications.}
\usage{
checkGeography(Directory, Filename)
}
\arguments{
\item{Directory}{A string identifying the path to the geographic
specifications file.}

\item{Filename}{A string identifying the name of the geographic
specifications file.}
}
\value{
A list having two components. The first component, 'Messages',
contains a string vector of error messages. It has a length of 0 if there are
no error messages. The second component, 'Update', is a list of components to
update in the model state file. The components of this list include: Geo, a
data frame that contains the geographic specifications; BzoneSpecified, a
logical identifying whether Bzones are specified; and CzoneSpecified, a
logical identifying whether Czones are specified.
}
\description{
\code{checkGeography} a visioneval framework control function that checks
geographic specifications file for model.
}
\details{
This function reads the file containing geographic specifications for the
model and checks the file entries to determine whether they are internally
consistent. This function is called by the readGeography function.
}
