% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{checkInputYearGeo}
\alias{checkInputYearGeo}
\title{Check years and geography of input file}
\usage{
checkInputYearGeo(Year_, Geo_, Group, Table)
}
\arguments{
\item{Year_}{the vector extract of the 'Year' column from the input data.}

\item{Geo_}{the vector extract of the 'Geo' column from the input data.}

\item{Group}{a string identifying the 'GROUP' specification for the data sets
contained in the input file.}

\item{Table}{a string identifying the 'TABLE' specification for the data sets
contained in the input file.}
}
\value{
A list containing the results of the check. The list has two
mandatory components and two optional components. 'CompleteInput' is a
logical that identifies whether records are present for all years and
geographic areas. 'DupInput' identifies where are any redundant year and
geography entries. If 'CompleteInput' is FALSE, the list contains a
'MissingInputs' component that is a string identifying the missing year and
geography records. If 'DupInput' is TRUE, the list contains a component that
is a string identifying the duplicated year and geography records.
}
\description{
\code{checkInputYearGeo} checks the 'Year' and 'Geo' columns of an input file
to determine whether they are complete and have no duplications.
}
\details{
This function checks the 'Year' and 'Geo' columns of an input file to
determine whether there are records for all run years specified for the
model and for all geographic areas for the level of geography. It also checks
for redundant year and geography entries.
}
