% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkIsElementOf}
\alias{checkIsElementOf}
\title{Check if data values are in a specified set of values}
\usage{
checkIsElementOf(Data_, SetElements_, DataName)
}
\arguments{
\item{Data_}{A vector of data of type integer, double, character, or logical.}

\item{SetElements_}{A vector of allowed values.}

\item{DataName}{A string identifying the field name of the data being
compared (used for composing message identifying non-compliant fields).}
}
\value{
A character vector of messages which identify the data field and the
  condition that is not met. A zero-length vector is returned if none of the
  conditions are met.
}
\description{
\code{checkIsElementOf} checks whether a data vector contains any elements
that are not in an allowed set of values.
}
\details{
This function is used to check whether categorical data values are consistent
with the defined set of allowed values.
}

