% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkMatchConditions}
\alias{checkMatchConditions}
\title{Check values with conditions.}
\usage{
checkMatchConditions(Data_, Conditions_, DataName, ConditionType)
}
\arguments{
\item{Data_}{A vector of data of type integer, double, character, or logical.}

\item{Conditions_}{A character vector of valid R comparison expressions or an
empty vector if there are no conditions.}

\item{DataName}{A string identifying the field name of the data being
compared (used for composing message identifying non-compliant fields).}

\item{ConditionType}{A string having a value of either "PROHIBIT" or
"UNLIKELY", the two data specifications which use conditions.}
}
\value{
A character vector of messages which identify the data field and the
  condition that is not met. A zero-length vector is returned if none of the
  conditions are met.
}
\description{
\code{checkMatchConditions} a visioneval framework control function that
checks whether a data vector contains any elements that match a set of
conditions.
}
\details{
This function checks whether any of the values in a data vector match one or
more conditions. The conditions are specified in a character vector where
each element is either "NA" (to match for the existence of NA values) or a
character representation of a valid R comparison expression for comparing
each element with a specified value (e.g. "< 0", "> 1", "!= 10"). This
function is used both for checking for the presence of prohibited values and
for the presence of unlikely values.
}
