% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkMatchType}
\alias{checkMatchType}
\title{Check data type}
\usage{
checkMatchType(Data_, Type, DataName)
}
\arguments{
\item{Data_}{A data vector.}

\item{Type}{A string identifying the specified data type.}

\item{DataName}{A string identifying the field name of the data being
compared (used for composing message identifying non-compliant fields).}
}
\value{
A list having 2 components, Errors and Warnings. If no error or
  warning is identified, both components will contain a zero-length character
  string. If either an error or warning is identified, the relevant component
  will contain a character string that identifies the data field and the type
  mismatch.
}
\description{
\code{checkMatchType} checks whether the data type of a data vector is
consistent with specifications.
}
\details{
This function checks whether the data type of a data vector is consistent
with a specified data type. An error message is generated if data can't be
coerced into the specified data type without the possibility of error or loss
of information (e.g. if a double is coerced to an integer). A warning message
is generated if the specified type is 'character' but the input data type is
'integer', 'double' or 'logical' since these can be coerced correctly, but
that may not be what is intended (e.g. zone names may be input as numbers).
}

