% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{checkModuleExists}
\alias{checkModuleExists}
\title{Check whether a module required to run a model is present}
\usage{
checkModuleExists(ModuleName, PackageName,
  InstalledPkgs_ = rownames(installed.packages()), CalledBy = NA)
}
\arguments{
\item{ModuleName}{A string identifying the module name.}

\item{PackageName}{A string identifying the package name.}

\item{InstalledPkgs_}{A string vector identifying the names of packages that
are installed.}

\item{CalledBy}{A string vector having two named elements. The value of the
'Module' element is the name of the calling module. The value of the
'Package' element is the name of the package that the calling module is in.}
}
\value{
TRUE if all packages and modules are present and FALSE if not.
}
\description{
\code{checkModuleExists} a visioneval framework control function that checks
whether a module required to run a model is present.
}
\details{
This function takes a specified module and package, checks whether the
package has been installed and whether the module is in the package. The
function returns an error message is the package is not installed or if
the module is not present in the package. If the module has been called by
another module the value of the 'CalledBy' argument will be used to identify
the calling module as well so that the user understands where the call is
coming from.
}
