% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{checkModuleOutputs}
\alias{checkModuleOutputs}
\title{Check module outputs for consistency with specifications}
\usage{
checkModuleOutputs(Data_ls, ModuleSpec_ls, ModuleName)
}
\arguments{
\item{Data_ls}{A list of all the datasets returned by a module in the
standard list form required by the VisionEval model system.}

\item{ModuleSpec_ls}{A list of module specifications in the standard list
form required by the VisionEval model system.}

\item{ModuleName}{A string identifying the name of the module.}
}
\value{
A character vector containing a list of error messages or having a
length of 0 if there are no error messages.
}
\description{
\code{checkModuleOutputs} a visioneval framework module developer function
that checks output list produced by a module for consistency with the
module's specifications.
}
\details{
This function is used to check whether the output list produced by a module
is consistent with the module's specifications. If there are any
specifications for creating tables, the function checks whether the output
list contains the table(s), if the LENGTH attribute of the table(s) are
present, and if the LENGTH attribute(s) are consistent with the length of the
datasets to be saved in the table(s). Each of the datasets in the output list
are checked against the specifications. These include checking that the
data type is consistent with the specified type and whether all values are
consistent with PROHIBIT and ISELEMENTOF conditions. For character types,
a check is made to ensure that a SIZE attribute exists and that the size
is sufficient to store all characters.
}
