% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{checkModulesExist}
\alias{checkModulesExist}
\title{Check whether required modules are present.}
\usage{
checkModulesExist(ModuleCalls_df)
}
\arguments{
\item{ModuleCalls_df}{A listing of modules called in the 'run_model.R' script
as created by the 'parseModelScript' function.}
}
\value{
TRUE if all packages and modules are present and FALSE if not.
}
\description{
\code{checkModulesExist}Function checks whether all required module
packages are installed and whether listed modules are present in the
packages.
}
\details{
This function takes a listing of module calls and checks whether all of the
listed packages are installed on the computer and whether the named modules
are in the packages. If modules are all present, then the function returns
TRUE, otherwise it identifies the missing packages and modules to the log and
returns FALSE.
}
