% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkModuleSpecs}
\alias{checkModuleSpecs}
\title{Checks all module specifications for completeness and for incorrect entries}
\usage{
checkModuleSpecs(Specs_ls, ModuleName)
}
\arguments{
\item{Specs_ls}{a module specifications list.}

\item{ModuleName}{a string identifying the name of the module. This is used in
the error messages to identify which module has errors.}
}
\value{
A vector containing messages identifying any errors that are found.
}
\description{
\code{checkModuleSpecs}checks all module specifications for
completeness and for proper values.
}
\details{
This function iterates through all the specifications for a module and
calls the checkSpec function to check each specification for completeness and
for proper values.
}
