% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkSpec}
\alias{checkSpec}
\title{Checks a module specifications for completeness and for incorrect entries}
\usage{
checkSpec(Spec_ls, SpecGroup, SpecNum)
}
\arguments{
\item{Spec_ls}{a list containing the specifications for a single item in
a module specifications list.}

\item{SpecGroup}{a string identifying the specifications group the
specification is in (e.g. RunBy, NewInpTable, NewSetTable, Inp, Get, Set).
This is used in the error messages to identify which specification has
errors.}

\item{SpecNum}{an integer identifying which specification in the
specifications group has errors.}
}
\value{
A vector containing messages identifying any errors that are found.
}
\description{
\code{checkSpec} a visioneval framework control function that checks a single
module specification for completeness and for proper values.
}
\details{
This function checks whether a single module specification (i.e. the
specification for a single dataset contains the minimum required
attributes and that the values of the attributes are correct.
}
