% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkSpecConsistency}
\alias{checkSpecConsistency}
\title{Check specification consistency}
\usage{
checkSpecConsistency(Spec_ls, DstoreAttr_)
}
\arguments{
\item{Spec_ls}{a list of data specifications consistent with a module "Get"
or "Set" specifications.}

\item{DstoreAttr_}{a named list where the components are the attributes of a
dataset.}
}
\value{
A list containing two components, Errors and Warnings. If no
  inconsistencies are found, both components will have zero-length character
  vectors. If there are one or more inconsistencies, then these components
  will hold vectors of error and warning messages. Mismatch between UNITS
  will produce a warning message. All other inconsistencies will produce
  error messages.
}
\description{
\code{checkSpecConsistency} checks whether the specifications for a dataset
are consistent with the data attributes in the datastore
}
\details{
This function compares the specifications for a dataset identified in a
module "Get" or "Set" are consistent with the attributes for that data in the
datastore.
}

