% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkSpecTypeUnits}
\alias{checkSpecTypeUnits}
\title{Checks the TYPE and UNITS and associated MULTIPLIER and YEAR attributes of a
Inp, Get, or Set specification for consistency.}
\usage{
checkSpecTypeUnits(Spec_ls, SpecGroup, SpecNum)
}
\arguments{
\item{Spec_ls}{a list for a single specification (e.g. a Get specification
for a dataset) that has been processed with the parseUnitsSpec function to
split the name, multiplier, and year elements of the UNITS specification.}

\item{SpecGroup}{a string identifying the group that this specification
comes from (e.g. Inp, Get, Set).}

\item{SpecNum}{a number identifying which specification in the order of the
SpecGroup. This is used to identify the subject specification if an error
is identified.}
}
\value{
A vector containing messages identifying any errors that are found.
}
\description{
\code{checkSpecTypeUnits}Checks correctness of TYPE, UNITS, MULTIPLIER and
YEAR attributes of a specification that has been processed with the
parseUnitsSpec function.
}
\details{
This function checks whether the TYPE and UNITS of a module's specification
contain errors. The check is done on a module specification in which the
module's UNITS attribute has been parsed by the parseUnitsSpec function to
split the name, multiplier, and years parts of the UNITS attribute. The TYPE
is checked against the types catalogued in the Types function. The units name
in the UNITS attribute is checked against the units names corresponding to
each type catalogued in the Types function. The MULTIPLIER is checked to
determine whether a value is a valid number, NA, or not a number (NaN). A NA
value means that no multiplier was specified (this is OK) a NaN value means
that a multiplier that is not a number was specified which is an error. The
YEAR attribute is checked to determine whether there is a proper
specification if the specified TYPE is currency. If the TYPE is currency, a
YEAR must be specified for Get and Set specifications.
}
