% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkTable}
\alias{checkTable}
\title{Check table existence}
\usage{
checkTable(Table, Year, DstoreListing_df, ThrowError = TRUE)
}
\arguments{
\item{Table}{a string representation of the table name.}

\item{Year}{a string or numeric representation of the year.}

\item{DstoreListing_df}{a dataframe which lists the contents of the datastore
as contained in the model state file.}

\item{ThrowError}{a logical identifying whether function execution should be
stopped if an error is found. The default value is TRUE.}
}
\value{
A string representation of the full path name for the table in the
datastore.
}
\description{
\code{checkTable} checks whether datastore contains a specified table
}
\details{
This function checks whether the datastore has a group for the specified
table and year. If so, it returns the full path name to the table.
}

