% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkTableExistence}
\alias{checkTableExistence}
\title{Check whether table exists in the datastore}
\usage{
checkTableExistence(Table, Group, DstoreListing_df)
}
\arguments{
\item{Table}{a string identifying the table.}

\item{Group}{a string or numeric representation of the group the table is a
part of.}

\item{DstoreListing_df}{a dataframe which lists the contents of the datastore
as contained in the model state file.}
}
\value{
A logical identifying whether a table is present in the datastore.
}
\description{
\code{checkTableExistence} checks whether a table is present in the
datastore.
}
\details{
This function checks whether a table is present in the datastore.
}

