% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkYear}
\alias{checkYear}
\title{Check year group existence}
\usage{
checkYear(Year, DstoreListing_df)
}
\arguments{
\item{Year}{A string or numeric representation of the year to check.}

\item{DstoreListing_df}{a dataframe which lists the contents of the datastore
as contained in the model state file.}
}
\value{
A string representation of the input year.
}
\description{
\code{checkYear} checks whether datastore has group for specified year
}
\details{
This function checks whether the datastore has a group for the argument year.
It also converts the year into a character type and returns the value.
}

