% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{createGeoIndex}
\alias{createGeoIndex}
\title{Create datastore index.}
\usage{
createGeoIndex(Table, Group, RunBy, Geo, GeoIndex_ls)
}
\arguments{
\item{Table}{A string identifying the name of the table the index is being
created for.}

\item{Group}{A string identifying the name of the group where the table is
located in the datastore.}

\item{RunBy}{A string identifying the level of geography the module is being
run at (e.g. Azone).}

\item{Geo}{A string identifying the geographic unit to create the index for
(e.g. the name of a particular Azone).}

\item{GeoIndex_ls}{a list of geographic indices used to determine the
positions to extract from a dataset corresponding to the specified
geography.}
}
\value{
A function that creates a vector of positions corresponding to the
  location of the supplied value in the index field.
}
\description{
\code{createIndex} a visioneval framework control function that creates an
index for reading or writing module data to the datastore.
}
\details{
This function creates indexing functions which return an index to positions
in datasets that correspond to positions in an index field of a table. For
example if the index field is 'Azone' in the 'Household' table, this function
will return a function that when provided the name of a particular Azone,
will return the positions corresponding to that Azone.
}
