% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{createGeoIndexList}
\alias{createGeoIndexList}
\title{Create a list of geographic indices for all tables in a datastore.}
\usage{
createGeoIndexList(Specs_ls, RunBy, RunYear)
}
\arguments{
\item{Specs_ls}{A 'Get' or 'Set' specifications list for a module.}

\item{RunBy}{The value of the RunBy specification for a module.}

\item{RunYear}{A string identifying the model year that is being run.}
}
\value{
A list that contains a component for each table identified in the
specifications in which each component includes all the geographic datasets
for the table represented by the component.
}
\description{
\code{createGeoIndexList} creates a list containing the geographic indices
for tables in the operating datastore for identified tables.
}
\details{
This function takes a 'Get' or 'Set' specifications list for a module and the
'RunBy' specification and returns a list which has a component for each table
identified in the specifications. Each component includes all geographic
datasets for the table.
}
