% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{createIndex}
\alias{createIndex}
\title{Create datastore index.}
\usage{
createIndex(Name, Table, Group)
}
\arguments{
\item{Name}{A string identifying the dataset the index is being created
for.}

\item{Table}{A string identifying the name of the table the index is
being created for.}

\item{Group}{A string identifying the group the table is located in.}
}
\value{
A function that creates a vector of positions corresponding to the
  location of the supplied value in the index field.
}
\description{
\code{createIndex} creates an index for reading or writing module data to the
datastore.
}
\details{
This function creates indexing functions which return an index to positions
in datasets that correspond to positions in an index field of a table. For
example if the index field is 'Azone' in the 'Household' table, this function
will return a function that when provided the name of a particular Azone,
will return the positions corresponding to that Azone.
}

