% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{deflateCurrency}
\alias{deflateCurrency}
\title{Convert currency values to different years.}
\usage{
deflateCurrency(Values_, FromYear, ToYear)
}
\arguments{
\item{Values_}{a numeric vector of values to convert from one currency year
to another.}

\item{FromYear}{a number or string identifying the currency year of the input
Values_.}

\item{ToYear}{a number or string identifying the currency year to convert the
Values_ to.}
}
\value{
A numeric vector of values corresponding the the input Values_ but
converted from the currency year identified in the FromYear argument to the
currency year identified in the ToYear argument. If either the FromYear or
the ToYear arguments is unaccounted for in the deflator series, the original
Values_ are returned with a Converted attribute of FALSE. Otherwise the
conversion is done and the Converted attribute of the returned values is TRUE.
}
\description{
\code{deflateCurrency} a visioneval framework control function that
converts currency values between different years of measure.
}
\details{
The visioneval framework stores all currency values in the base year real
currency (e.g. dollar) values. However, currency inputs may be in different
nominal year currency. Also modules may be estimated using different nominal
year currency data. For example, the original vehicle travel model in
GreenSTEP used 2001 NHTS data while the newer model uses 2009 NHTS data. The
framework enables model uses to specify the currency year in the field name
of an input file that contains currency data. Likewise, the currency year can
be encoded in the UNIT attributes for a modules Get and Set specifications.
The framework converts dollars to and from specified currency year values and
base year real dollar values. The model uses a set of deflator values that
the user inputs for the region to make the adjustments. These values are
stored in the model state list.
}
