% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{documentModule}
\alias{documentModule}
\title{Produces markdown documentation for a module}
\usage{
documentModule(ModuleName)
}
\arguments{
\item{ModuleName}{A string identifying the name of the module
(e.g. 'CalculateHouseholdDvmt')}
}
\value{
None. The function has the side effects of creating a 'vignettes'
directory if one does not exist, copying identified 'fig' resources to the
'vignettes' directory, and saving the markdown documentation file to the
'vignettes' directory. The markdown file is named with the module name and
has a 'md' suffix.
}
\description{
\code{documentModule} a visioneval framework module developer function
that creates a vignettes directory if one does not exist and produces
module documentation in markdown format which is saved in the vignettes
directory.
}
\details{
This function produces documentation for a module in markdown format. A
'vignettes' directory is created if it does not exist and the markdown file
and any associated resources such as image files are saved in that directory.
The function is meant to be called within and at the end of the module
script. The documentation is created from a commented block within the
module script which is enclosed by the opening tag, <doc>, and the closing
tag, </doc>. (Note, these tags must be commented along with all the other
text in the block). This commented block may also include tags which identify
resources to include within the documentation. These tags identify the
type of resource and the name of the resource which is located in the 'data'
directory. A colon (:) is used to separate the resource type and resource
name identifiers. For example:
<txt:DvmtModel_ls$EstimationStats$NonMetroZeroDvmt_GLM$Summary>
is a tag which will insert text which is located in a component of the
DvmtModel_ls list that is saved as an rdata file in the 'data' directory
(i.e. data/DvmtModel_ls.rda). The following 3 resource types are recognized:
* txt - a vector of strings which are inserted as lines of text in a code block
* fig - a png file which is inserted as an image
* tab - a matrix or data frame which is inserted as a table
The function also reads in the module specifications and creates
tables that document user input files, data the module gets from the
datastore, and the data the module produces that is saved in the datastore.
This function is intended to be called in the R script which defines the
module. It is placed near the end of the script (after the portions of the
script which estimate module parameters and define the module specifications)
so that it is run when the package is built. It may not properly in other
contexts.
}
