% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{doProcessGetSpec}
\alias{doProcessGetSpec}
\title{Filters Get specifications list based on OPTIONAL specification attributes.}
\usage{
doProcessGetSpec(GetSpecs_ls, ProcessedInpSpecs_ls, DstoreListing_df, RunYears_)
}
\arguments{
\item{GetSpecs_ls}{A standard specifications list for Get specifications.}

\item{ProcessedInpSpecs_ls}{A processed specifications of Inp specifications
that has been processed by the 'doProcessInpSpec' to remove all OPTIONAL
specifications that will not be processed.}

\item{DstoreListing_df}{A standard datastore listing dataframe.}

\item{RunYears_}{A string vector of the model run years.}
}
\value{
A list containing the Get specification components that meet the
criteria of either not being optional or being optional and the specified
input file is present.
}
\description{
\code{doProcessGetSpec}Filters out Get specifications whose OPTIONAL
specification attribute is TRUE but the specified dataset is not included in
the corresponding Inp specifications nor is it present in the datastore.
}
\details{
A Get specification component may have an OPTIONAL specification whose value
is TRUE. If so, and if the specified dataset will be created by processing
the corresponding Inp specifications or if the dataset is already present in
the datastore then the get specification needs to be processed. This function
checks whether the OPTIONAL specification is present, whether its value is
TRUE, and whether the dataset will be created by processing the corresponding
Inp specifications or if it exists in the datastore for all model run years.
If all of these are true, then the get specification needs to be processed.
The get specification also needs to be processed if it is not optional. A
specification is not optional if the OPTIONAL attribute is not present or if
it is present and the value is not TRUE. The function returns a list of all
the Get specifications that meet these criteria.
}
