% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{doProcessSetSpec}
\alias{doProcessSetSpec}
\title{Filters Set specifications list based on OPTIONAL specification attributes.}
\usage{
doProcessSetSpec(SetSpecs_ls, ProcessedGetSpecs_ls)
}
\arguments{
\item{SetSpecs_ls}{A standard specifications list for Set specifications.}

\item{ProcessedGetSpecs_ls}{A processed specifications of Get specifications
that has been processed by the 'doProcessGetSpec' to remove all OPTIONAL
specifications that will not be processed.}
}
\value{
A list containing the Set specification components that meet the
criteria of either not being optional or being optional and the specified
input file is present.
}
\description{
\code{doProcessSetSpec} a visioneval framework control function that filters
out Set specifications whose OPTIONAL specification attribute is TRUE but the
specified dataset is not included in the corresponding Get specifications.
}
\details{
A Set specification component may have an OPTIONAL specification whose value
is TRUE. If so, and if the specified dataset is included in the corresponding
Get specifications then the Set specification component needs to be
processed. This function checks whether the OPTIONAL specification is
present, whether its value is TRUE, and whether the dataset will be created
by processing the corresponding Get specifications. If all of these are true,
then the Set specification needs to be processed. The Set specification also
needs to be processed if it is not optional. A specification is not optional
if the OPTIONAL attribute is not present or if it is present and the value is
not TRUE. The function returns a list of all the Set specifications that meet
these criteria. It is important to note that optional Set specifications only
exist to enable modules to do additional processing of input data. For
example, to adjust proportions so that they exactly equal 1. In this general
use case, optional input datasets are loaded, then they are retrieved from
the datastore, processed, and then resaved to the datastore.
}
