% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{expandSpec}
\alias{expandSpec}
\title{Expand a Inp, Get, or Set specification so that is can be used by other
functions to process inputs and to read from or write to the datastore.}
\usage{
expandSpec(SpecToExpand_ls, ComponentName)
}
\arguments{
\item{SpecToExpand_ls}{A standard specifications list for a specification
whose NAME attribute has multiple values.}

\item{ComponentName}{A string that is the name of the specifications
that the specification is a part of (e.g. "Inp", "Get", "Set").}
}
\value{
A list of standard specifications lists which has a component for
each value in the NAME attribute of the input specifications list.
}
\description{
\code{expandSpec} a visioneval framework control function that takes a Inp,
Get, or Set specification and processes it to be in a form that can be used
by other functions which use the specification in processing inputs or
reading from or writing to the datastore. The parseUnitsSpec function is
called to parse the UNITS attribute to extract name, multiplier, and year
values. When the specification has multiple values for the NAME attribute,
the function creates a specification for each name value.
}
\details{
The VisionEval design allows module developers to assign multiple values to
the NAME attributes of a Inp, Get, or Set specification where the other
attributes for those named datasets (or fields) are the same. This greatly
reduces duplication and the potential for error in writing module
specifications. However, other functions that check or use the specifications
are not capable of handling specifications which have NAME attributes
containing multiple values. This function expands a specification with
multiple values for a  NAME attribute into multiple specifications, each with
a single value for the NAME attribute. In addition, the function calls the
parseUnitsSpec function to extract multiplier and year information from the
value of the UNITS attribute. See that function for details.
}
