% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{findSpec}
\alias{findSpec}
\title{Find the full specification corresponding to a defined NAME, TABLE, and GROUP}
\usage{
findSpec(Specs_ls, Name, Table, Group)
}
\arguments{
\item{Specs_ls}{a standard specifications list for 'Inp', 'Get', or 'Set'}

\item{Name}{a string for the name of the dataset}

\item{Table}{a string for the table that the dataset resides in}

\item{Group}{a string for the generic group that the table resides in}
}
\value{
A list containing the full specifications for the dataset
}
\description{
\code{findSpec} returns the full dataset specification for defined NAME,
TABLE, and GROUP.
}
\details{
This function finds and returns the full specification from a specifications
list whose NAME, TABLE and GROUP values correspond to the Name, Table, and
Group argument values. The specifications list must be in standard format and
must be for only 'Inp', 'Get', or 'Set' specifications.
}
