% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getDatasetAttr}
\alias{getDatasetAttr}
\title{Get attributes of a dataset}
\usage{
getDatasetAttr(Name, Table, Group, DstoreListing_df)
}
\arguments{
\item{Name}{a string identifying the dataset name.}

\item{Table}{a string identifying the table the dataset is a part of.}

\item{Group}{a string or numeric representation of the group the table is a
part of.}

\item{DstoreListing_df}{a dataframe which lists the contents of the datastore
as contained in the model state file.}
}
\value{
A named list of the dataset attributes.
}
\description{
\code{getDatasetAttr} a visioneval framework control function that retrieves
the attributes for a dataset in the datastore.
}
\details{
This function extracts the listed attributes for a specific dataset from the
datastore listing.
}
