% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{getFromDatastore}
\alias{getFromDatastore}
\title{Retrieve data identified in 'Get' specifications from datastore}
\usage{
getFromDatastore(ModuleSpec_ls, RunYear, Geo = NULL, GeoIndex_ls = NULL)
}
\arguments{
\item{ModuleSpec_ls}{a list of module specifications that is consistent with
the VisionEval requirements}

\item{RunYear}{a string identifying the model year being run. The default is
the Year object in the global workspace.}

\item{Geo}{a string identifying the name of the geographic area to get the
data for. For example, if the module is specified to be run by Azone, then
Geo would be the name of a particular Azone.}

\item{GeoIndex_ls}{a list of geographic indices used to determine the
positions to extract from a dataset corresponding to the specified geography.}
}
\value{
A list containing all the data sets specified in the module's
'Get' specifications for the identified geographic area.
}
\description{
\code{getFromDatastore} a visioneval framework control function that
retrieves datasets identified in a module's 'Get' specifications from the
datastore.
}
\details{
This function retrieves from the datastore all of the data sets identified in
a module's 'Get' specifications. If the module's specifications include the
name of a geographic area, then the function will retrieve the data for that
geographic area.
}
