% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{getModuleSpecs}
\alias{getModuleSpecs}
\title{Retrieve module specifications from a package}
\usage{
getModuleSpecs(ModuleName, PackageName)
}
\arguments{
\item{ModuleName}{A string identifying the name of the module.}

\item{PackageName}{A string identifying the name of the package that the
module is in.}
}
\value{
A specifications list that is the same as the specifications list
defined for the module in the package.
}
\description{
\code{getModuleSpecs}Function retrieves the specifications list for a module
and returns the specifications list.
}
\details{
This function loads the specifications for a module in a package. It returns
the specifications list.
}
