% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getTableLength}
\alias{getTableLength}
\title{Get the length of a table in the datastore}
\usage{
getTableLength(Table, Group, DstoreListing_df)
}
\arguments{
\item{Table}{a string identifying the table.}

\item{Group}{a string or numeric representation of the group the table is a
part of.}

\item{DstoreListing_df}{a dataframe which lists the contents of the datastore
as contained in the model state file.}
}
\value{
A number that table length.
}
\description{
\code{getTableLength} retrieves the LENGTH attribute for a table in the
datastore.
}
\details{
This function extracts the LENGTH attribute for a table in the datastore.
}

