% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{getUnits}
\alias{getUnits}
\title{Retrieve default units for model}
\usage{
getUnits(Type)
}
\arguments{
\item{Type}{A string identifying the complex data type.}
}
\value{
A string identifying the default units for the complex data type.
}
\description{
\code{getUnits} retrieves the default model units for a complex data type.
}
\details{
This is a convenience function to make it easier to retrieve the default
units for a complex data type (e.g. distance, volume, speed). The default
units are the units used to store the complex data type in the datastore.
}
