% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{getYears}
\alias{getYears}
\title{Retrieve years}
\usage{
getYears()
}
\value{
A character vector of the model run years.
}
\description{
\code{getYears} reads the Years component from the the model state file.
}
\details{
This is a convenience function to make it easier to retrieve the Years
component of the model state file. If the Years component includes the base
year then order the Years component so that it is first. This ordering is
important because some modules calculate future year values by pivoting off
of base year values so the base year must be run first.
}
