% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{hasErrors}
\alias{hasErrors}
\title{Return module errors}
\usage{
hasErrors(ModuleCheck_ls)
}
\arguments{
\item{ModuleCheck_ls}{A list that is returned from the "testModule" function.}
}
\value{
None. The function stops script execution if there are any errors and
the message contains a list of the errors.
}
\description{
\code{hasErrors} checks whether the return value from the testModule function
contains any error messages and returns a message stating them.
}
\details{
This function checks whether the results from testing a module contain any
errors. The function stops script execution if there are any errors and the
message contains a list of the errors.
}
