% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initDataList}
\alias{initDataList}
\title{Initialize a list for data transferred to and from datastore}
\usage{
initDataList()
}
\value{
A list that has 3 named list components: Global, Year, BaseYear
}
\description{
\code{initDataList} creates a list to be used for transferring data to and
from the datastore.
}
\details{
This function initializes a list to store data that is transferred from
the datastore to a module or returned from a module to be saved in the
datastore. The list has 3 named components (Global, Year, and BaseYear). This
is the standard structure for data being passed to and from a module and the
datastore.
}
