% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{initDataset}
\alias{initDataset}
\title{Initialize dataset in datastore table.}
\usage{
initDataset(Spec_ls, Group)
}
\arguments{
\item{Spec_ls}{a list containing the standard module 'Set' specifications
described in the model system design documentation.}

\item{Group}{a string representation of the name of the group the table is to
be created in.}
}
\value{
TRUE if dataset is successfully initialized. If the dataset already
exists the function throws an error and writes an error message to the log.
Updates the model state file.
}
\description{
\code{initDataset} initializes a dataset in a table.
}
\details{
This function initializes a dataset which must be done before data can be
stored. Initialization establishes a name for the dataset and key attributes
associated with the data. Function calls initTable if the table the
dataset is to be placed in has not been initialized.
}
