% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initDatasetH5}
\alias{initDatasetH5}
\title{Initialize dataset in an HDF5 (H5) type datastore table.}
\usage{
initDatasetH5(Spec_ls, Group)
}
\arguments{
\item{Spec_ls}{a list containing the standard module 'Set' specifications
described in the model system design documentation.}

\item{Group}{a string representation of the name of the group the table is to
be created in.}
}
\value{
TRUE if dataset is successfully initialized. If the dataset already
exists the function throws an error and writes an error message to the log.
Updates the model state file.
}
\description{
\code{initDatasetH5} a visioneval framework datastore connection function
that initializes a dataset in an HDF5 (H5) type datastore table.
}
\details{
This function initializes a dataset in an HDF5 (H5) type datastore table.
}
