% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initDatasetRD}
\alias{initDatasetRD}
\title{Initialize dataset in an RData (RD) type datastore table.}
\usage{
initDatasetRD(Spec_ls, Group)
}
\arguments{
\item{Spec_ls}{a list containing the standard module specifications
described in the model system design documentation.}

\item{Group}{a string representation of the name of the top-level
subdirectory the table is to be created in (i.e. either 'Global' or the name
of the year).}
}
\value{
TRUE if dataset is successfully initialized. If the identified table
does not exist, the function throws an error.
}
\description{
\code{initDatasetRD} initializes a dataset in an RData (RD) type datastore
table.
}
\details{
This function initializes a dataset in an RData (RD) type datastore table.
}
