% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{initDatastore}
\alias{initDatastore}
\title{Initialize Datastore.}
\usage{
initDatastore()
}
\value{
TRUE if datastore initialization is successful. Calls the
listDatastore function which adds a listing of the datastore contents to the
model state file.
}
\description{
\code{initDatastore} creates datastore with starting structure.
}
\details{
This function creates the datastore file for the model run with the initial
structure. The file is created in the working directory and is named
'datastore.h5'. If this file already exists it is deleted and a new
initialized file is created. Groups are created in the file; one for each
model run year and one named 'Global' for storing global values.
}
