% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visioneval.R
\name{initializeModel}
\alias{initializeModel}
\title{Initialize model.}
\usage{
initializeModel(Dir = "defs", ParamFile = "parameters.json",
  GeoFile = "geo.csv")
}
\arguments{
\item{Dir}{A string identifying the relative or absolute path to the
directory where the parameter and geography definition files are located.
The default value is "defs".}

\item{ParamFile}{A string identifying the name of a JSON-formatted text file
that contains global parameters needed by the framework in order to set up
the working environment. The default value is "parameters.json".}

\item{GeoFile}{A string identifying the name of a text file in
comma-separated values format that contains the geographic specifications
for the model. The default value is "geo.csv".}
}
\value{
None
}
\description{
\code{initializeModel}Function initializes a VisionEval model given the
location of files that establish model parameters and geography.
}
\details{
This function does several things to initialize the model environment and
datastore including: initializing a file that is used to keep track of the
state of key model run variables and the datastore, initializes a log to
which messages are written, creates the datastore and initializes its
structure, reads in and checks the geographic specifications, and initializes
the geography in the datastore. <TO DO: ENABLE LOADING OF EXISTING DATASTORE,
CHECK WHETHER ALL SPECIFIED PACKAGES EXIST, WHETHER SPECIFIED INPUT FILES
EXIST AND CONTENTS ARE CORRECT, AND WHETHER ALL DATA DEPENDENCIES WIL BE
SATISFIED.>
}

