% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initLog}
\alias{initLog}
\title{Initialize run log.}
\usage{
initLog(Suffix = NULL)
}
\arguments{
\item{Suffix}{A character string appended to the file name for the log file.
For example, if the suffix is 'CreateHouseholds', the log file is named
'Log_CreateHouseholds.txt'. The default value is NULL in which case the
suffix is the date and time.}
}
\value{
TRUE if the log is created successfully. It creates a log file in the
  working directory and identifies the name of the log file in the
  model state file.
}
\description{
\code{initLog} a visioneval framework control function that creates a log
(text file) that stores messages generated during a model run.
}
\details{
This function creates a log file that is a text file which stores messages
generated during a model run. The name of the log is 'Log <date> <time>'
where '<date>' is the initialization date and '<time>' is the initialization
time. The log is initialized with the scenario name, scenario description and
the date and time of initialization.
}
