% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initModelStateFile}
\alias{initModelStateFile}
\title{Initialize model state.}
\usage{
initModelStateFile(Dir = "defs", ParamFile = "parameters.json")
}
\arguments{
\item{Dir}{A string identifying the name of the directory where the global
parameters file is located. The default value is "defs".}

\item{ParamFile}{A string identifying the name of the global parameters file.
The default value is "parameters.json".}
}
\value{
TRUE if the model state file is created. It creates the model state
file and loads parameters recorded in the 'parameters.json' file into the
model state file.
}
\description{
\code{initModelStateFile} loads model run parameters into the model state
file that is used to keep track of the model state.
}
\details{
This function creates the model state file and loads model run parameters
recorded in the 'parameters.json' file into the model state file.
}

