% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initTable}
\alias{initTable}
\title{Initialize table in datastore.}
\usage{
initTable(Table, Group, Length)
}
\arguments{
\item{Table}{a string identifying the name of the table to initialize.}

\item{Group}{a string representation of the name of the top-level
subdirectory the table is to be created in (i.e. either 'Global' or the name
of the year).}

\item{Length}{a number identifying the table length.}
}
\value{
The value TRUE is returned if the function is successful at creating
  the table. In addition, the listDatastore function is run to update the
  inventory in the model state file. The function stops if the group in which
  the table is to be placed does not exist in the datastore and a message is
  written to the log.
}
\description{
\code{initDatastoreTable} initializes a table in the datastore.
}
\details{
A table in the datastore is a directory which contains one or more R binary
files. Each file is a dataset that is a vector or list. The length of each
vector or list must be the same. Thus a table is much like an R data frame
which is a list containing equal length vectors that may have different
types. If the dataset is a list, then it is like a list column in a
dataframe. A table is initialized in the datastore by creating a directory
whose name is the table name. The DatastoreListing_ls is updated to include
the 'group', 'name', 'groupname', and 'attributes' information for the table.
The 'attributes' listing for the table includes one element (LENGTH) that is
the length that each dataset in the table must have.
}
