% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initTableH5}
\alias{initTableH5}
\title{Initialize table in an HDF5 (H5) type datastore.}
\usage{
initTableH5(Table, Group, Length)
}
\arguments{
\item{Table}{a string identifying the name of the table to initialize.}

\item{Group}{a string representation of the name of the group the table is to
be created in.}

\item{Length}{a number identifying the table length.}
}
\value{
The value TRUE is returned if the function is successful at creating
  the table. In addition, the listDatastore function is run to update the
  inventory in the model state file. The function stops if the group in which
  the table is to be placed does not exist in the datastore and a message is
  written to the log.
}
\description{
\code{initTableH5} a visioneval framework datastore connection function that
initializes a table in an HDF5 (H5) type datastore.
}
\details{
This function initializes a table in an HDF5 (H5) type datastore.
}
