% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{inputsToDatastore}
\alias{inputsToDatastore}
\title{Write the datasets in a list of module inputs that have been processed to the
datastore.}
\usage{
inputsToDatastore(Inputs_ls, ModuleSpec_ls, ModuleName)
}
\arguments{
\item{Inputs_ls}{a list processes module inputs as created by the
'processModuleInputs' function.}

\item{ModuleSpec_ls}{a list of module specifications that is consistent with
the VisionEval requirements.}

\item{ModuleName}{a string identifying the name of the module (used to
document the dataset in the datastore).}
}
\value{
A logical indicating successful completion. Most of the outputs of
the function are the side effects of writing data to the datastore.
}
\description{
\code{inputsToDatastore} takes a list of processed module input files and
writes the datasets to the datastore.
}
\details{
This function takes a processed list of input datasets specified by a module
created by the application of the 'processModuleInputs' function and writes
the datasets in the list to the datastore.
}
