% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{item}
\alias{item}
\title{Alias for list function.}
\usage{
item()
}
\value{
a list.
}
\description{
\code{item} is an alias for the list function whose purpose is to make
module specifications easier to read.
}
\details{
This function defines an alternate name for list. It is used in module
specifications to identify data items in the Inp, Get, and Set portions of
the specifications.
}
