% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{items}
\alias{items}
\title{Alias for list function.}
\usage{
items()
}
\value{
a list.
}
\description{
\code{items} a visioneval framework \strong{module developer} function that is
an alias for the list function whose purpose is to make module specifications
easier to read.
}
\details{
This function defines an alternate name for list. It is used in module
specifications to identify a group of data items in the Inp, Get, and Set
portions of the specifications.
}
