% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{listDatastore}
\alias{listDatastore}
\title{List datastore contents.}
\usage{
listDatastore()
}
\value{
TRUE if the listing is successfully read from the datastore and
written to the model state file.
}
\description{
\code{listDatastore} lists the contents of a datastore.
}
\details{
This function lists the contents of a datastore including identifying all
groups, tables, and datasets. It also lists the attributes associated with
each table and dataset. The listing is stored in the model state file
(ModelState.Rda) as the "Datastore" component. This function is run whenever
new groups, tables, or datasets are created in the datastore to always keep
the listing in the model state file current.
}
