% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{listDatastoreH5}
\alias{listDatastoreH5}
\title{List datastore contents for an HDF5 (H5) type datastore.}
\usage{
listDatastoreH5()
}
\value{
TRUE if the listing is successfully read from the datastore and
written to the model state file.
}
\description{
\code{listDatastoreH5} lists the contents of an HDF5 (H5) type datastore.
}
\details{
This function lists the contents of a datastore for an HDF5 (H5) type
datastore.
}
