% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{listDatastoreRD}
\alias{listDatastoreRD}
\title{List datastore contents for an RData (RD) type datastore.}
\usage{
listDatastoreRD(DataListing_ls = NULL)
}
\arguments{
\item{DataListing_ls}{a list containing named elements describing a new data
item being added to the datastore listing and the model state file. The list
components are:
group - the name of the group (path) the item is being added to;
name - the name of the data item (directory or dataset);
groupname - the full path to the data item;
attributes - a list containing the named attributes of the data item.}
}
\value{
TRUE if the listing is successfully read from the datastore and
written to the model state file.
}
\description{
\code{listDatastoreRD} a visioneval framework datastore connection function
that lists the contents of an RData (RD) type datastore.
}
\details{
This function lists the contents of a datastore for an RData (RD) type
datastore.
}
