% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{loadDatastore}
\alias{loadDatastore}
\title{Load saved datastore.}
\usage{
loadDatastore(FileToLoad, GeoFile)
}
\arguments{
\item{FileToLoad}{A string identifying the full path name to the saved
datastore. Path name can either be relative to the working directory or
absolute.}

\item{GeoFile}{A string identifying the name of the geography definition file
(see 'readGeography' function) that is consistent with the saved datastore.
The geography definition file must be located in the 'defs' directory.}
}
\value{
TRUE if the datastore is loaded. It copies the saved datastore to
  working directory as 'datastore.h5'. If a 'datastore.h5' file already
  exists, it first renames that file as 'archive-datastore.h5'. The function
  updates information in the model state file regarding the model geography
  and the contents of the loaded datastore. If the stored file does not exist
  an error is thrown.
}
\description{
\code{loadDatastore} copy an existing saved datastore and write information
to run environment.
}
\details{
This function copies a saved datastore as the working datastore attributes
the global list with related geographic information. This function enables
scenario variants to be built from a constant set of starting conditions.
}

