% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{loadModelParameters}
\alias{loadModelParameters}
\title{Load model global parameters file into datastore.}
\usage{
loadModelParameters(ModelParamFile = "model_parameters.json")
}
\arguments{
\item{ModelParamFile}{A string identifying the name of the parameter file.
The default value is 'model_parameters.json'.}
}
\value{
The function returns TRUE if the model parameters file exists and
its values are sucessfully written to the datastore.
}
\description{
\code{loadModelParameters} reads the 'model_parameters.json' file and
stores the contents in the 'Global/Model' group of the datastore.
}
\details{
This function reads the 'model_parameters.json' file in the 'defs' directory
which contains parameters specific to a model rather than to a module. These
area parameters that may be used by any module. Parameters are specified by
name, value, and data type. The function creates a 'Model' group in the
'Global' group and stores the values of the appropriate type in the 'Model'
group.
}

