% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{makeModelFormulaString}
\alias{makeModelFormulaString}
\title{Makes a string representation of a model equation.}
\usage{
makeModelFormulaString(EstimatedModel)
}
\arguments{
\item{EstimatedModel}{the return value of the 'lm' or 'glm' functions.}
}
\value{
a string expression of the model equation.
}
\description{
\code{makeModelFormulaString} a visioneval framework module developer
function that creates a string equivalent of a model equation.
}
\details{
The return values of model estimation functions such as 'lm' and 'glm'
contain a large amount of information in addition to the parameter estimates
for the specified model. This is particularly the case when the estimation
dataset is large. Most of this information is not needed to apply the model
and including it can add substantially to the size of a package that includes
several estimated models. All that is really needed to implement an estimated
model is an equation of the model terms and estimated coefficients. This
function creates a string representation of the model equation.
}
