% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{parseModelScript}
\alias{parseModelScript}
\title{Parse model script.}
\usage{
parseModelScript(FilePath = "run_model.R")
}
\arguments{
\item{Path}{A string identifying the relative or absolute path to the
model run script is located.}
}
\value{
A data frame containing information on the calls to 'runModule' in the
order of the calls. Each row represents a module call in order. The columns
identify the 'ModuleName', the 'PackageName', and the 'TimeFrame'.
}
\description{
\code{parseModel}Function reads and parses the model script to identify the
sequence of module calls and the associated call arguments.
}
\details{
This function reads in the model run script and parses the script to
identify the sequence of module calls. It extracts each call to 'runModule'
and identifies the values assigned to the function arguments. It creates a
list of the calls with their arguments in the order of the calls in the
script.
}

