% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{parseUnitsSpec}
\alias{parseUnitsSpec}
\title{Parse units specification into components and add to specifications list.}
\usage{
parseUnitsSpec(Spec_ls)
}
\arguments{
\item{Spec_ls}{A standard specifications list for a Inp, Get, or Set item.}
}
\value{
a list that is a standard specifications list with the addition of
a MULTIPLIER component and a YEAR component as well as a modification of the
UNIT component. The MULTIPLIER component can have the value of NA, a number,
or NaN. The value is NA if the multiplier is missing. It is a number if the
multiplier is a valid number. The value is NaN if the multiplier is not a
valid number. The YEAR component is a character string that is a 4-digit
representation of a year or NA if the component is missing or not a proper
year. The UNITS component is modified to only be the units name.
}
\description{
\code{parseUnitsSpec} a visioneval framework control function that parses the
UNITS attribute of a standard Inp, Get, or Set specification for a dataset to
identify the units name, multiplier, and year for currency data. Returns a
modified specifications list whose UNITS value is only the units name, and
includes a MULTIPLIER attribute and YEAR attribute.
}
\details{
The UNITS component of a specifications list can encode information in
addition to the units name. This includes a value units multiplier and in
the case of currency values the year for the currency measurement. The
multiplier element can only be expressed in scientific notation where the
number before the 'e' can only be 1.
}
