% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{processEstimationInputs}
\alias{processEstimationInputs}
\title{Load estimation data}
\usage{
processEstimationInputs(Inp_ls, FileName, ModuleName)
}
\arguments{
\item{Inp_ls}{A list that describes the specifications for the estimation
file. This list must meet the framework standards for specification
description.}

\item{FileName}{A string identifying the file name. This is the file name
without any path information. The file must located in the "inst/extdata"
directory of the package.}

\item{ModuleName}{A string identifying the name of the module the estimation
data is being used in.}
}
\value{
A data frame containing the estimation data according to
specifications with data types consistent with specifications and columns
not specified removed. Execution stops if any errors are found. Error
messages are printed to the console. Warnings are also printed to the console.
}
\description{
\code{processEstimationInputs} a visioneval framework module developer
function that checks whether specified model estimation data meets
specifications and returns the data in a data frame.
}
\details{
This function is used to check whether a specified CSV-formatted data file
used in model estimation is correctly formatted and contains acceptable
values for all the datasets contained within. The function checks whether the
specified file exists in the "inst/extdata" directory. If the file does not
exist, the function stops and transmits a standard error message that the
file does not exist. If the file does exist, the function reads the file into
the data frame and then checks whether it contains the specified columns and
that the data meets all specifications. If any of the specifications are not
met, the function stops and transmits an error message. If there are no
data errors the function returns a data frame containing the data in the
file.
}
