% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{processModuleInputs}
\alias{processModuleInputs}
\title{Process module input files}
\usage{
processModuleInputs(ModuleSpec_ls, ModuleName, Dir = "inputs")
}
\arguments{
\item{ModuleSpec_ls}{a list of module specifications that is consistent with
the VisionEval requirements.}

\item{ModuleName}{a string identifying the name of the module (used to document
module in error messages).}

\item{Dir}{a string identifying the relative path to the directory where the
model inputs are contained.}
}
\value{
A list containing the results of the input processing. The list has
two components. The first (Errors) is a vector of identified file and data
errors. The second (Data) is a list containing the data in the input files
organized in the standard format for data exchange with the datastore.
}
\description{
\code{processModuleInputs} processes input files identified in a module's
'Inp' specifications in preparation for saving in the datastore.
}
\details{
This function processes the input files identified in a module's 'Inp'
specifications in preparation for saving the data in the datastore. Several
processes are carried out. The existence of each specified input file is
checked. Files that are not global, are checked to determine that they have
'Year' and 'Geo' columns. The entries in the 'Year' and 'Geo' columns are
checked to make sure they are complete and there are no duplicates. The data
in each column are checked against specifications to determine conformance.
The function returns a list which contains a list of error messages and a
list of the data inputs. The function also writes error messages and warnings
to the log file.
}
