% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{readDatastoreTables}
\alias{readDatastoreTables}
\title{Read multiple datasets from multiple tables in datastores}
\usage{
readDatastoreTables(Tables_ls, Group, DstoreLocs_, DstoreType)
}
\arguments{
\item{Tables_ls}{a named list where the name of each component is the name of
a table in a datastore group and the value is a string vector of the names
of the datasets to be retrieved.}

\item{Group}{a string that is the name of the group to retrieve the table
datasets from.}

\item{DstoreLocs_}{a string vector identifying the paths to all of the
datastores to extract the datasets from. Each entry must be the full relative
path to a datastore (e.g. 'tests/Datastore').}

\item{DstoreType}{a string identifying the type of datastore
(e.g. 'RD', 'H5'). Note}
}
\value{
A named list having two components. The 'Data' component is a list
containing the datasets from the datastores where the name of each component
of the list is the name of a table from which identified datasets are
retrieved and the value is a data frame containing the identified datasets.
The 'Missing' component is a list which identifies the datasets that are
missing in each table.
}
\description{
\code{readDatastoreTables} reads datasets from one or more tables in a
specified group in one or more datastores
}
\details{
This function can read multiple datasets in one or more tables in a group.
More than one datastore my be specified so that if datastore references are
used in a model run, datasets from the referenced datastores may be queried
as well. Note that the capability for querying multiple datastores is only
for the purpose of querying datastores for a single model scenario. This
capability should not be used to compare multiple scenarios. The function
does not segregate datasets by datastore. Attempting to use this function to
compare multiple scenarios could produce unpredictable results.
}
