% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{readFromTableH5}
\alias{readFromTableH5}
\title{Read from an HDF5 (H5) type datastore table.}
\usage{
readFromTableH5(Name, Table, Group, File = NULL, Index = NULL)
}
\arguments{
\item{Name}{A string identifying the name of the dataset to be read from.}

\item{Table}{A string identifying the complete name of the table where the
dataset is located.}

\item{Group}{a string representation of the name of the datastore group the
data is to be read from.}

\item{File}{a string representation of the file path of the datastore}

\item{Index}{A numeric vector identifying the positions the data is to be
written to. NULL if the entire dataset is to be read.}
}
\value{
A vector of the same type stored in the datastore and specified in
  the TYPE attribute.
}
\description{
\code{readFromTableH5} a visioneval framework datastore connection function
that reads a dataset from an HDF5 (H5) type datastore table.
}
\details{
This function reads a dataset from an HDF5 (H5) type datastore table.
}
