% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{readGeography}
\alias{readGeography}
\title{Read geographic specifications.}
\usage{
readGeography(Dir = "defs", GeoFile = "geo.csv")
}
\arguments{
\item{Dir}{A string identifying the path to the geographic
specifications file. Note: don't include the final separator in the
path name 'e.g. not defs/'.}

\item{GeoFile}{A string identifying the name of the geographic
specifications file. This is a csv-formatted text file which contains
columns named 'Azone', 'Bzone', 'Czone', and 'Marea'. The 'Azone' column
must have zone names in all rows. The 'Bzone' and 'Czone' columns can be
unspecified (NA in all rows) or may have have unique names in every row.
The 'Marea' column (referring to metropolitan areas) identifies
metropolitan areas corresponding to the most detailed level of specified
geography (or 'None' no metropolitan area occupies any portion of the
zone.}
}
\value{
The value TRUE is returned if the function is successful at reading
  the file and the specifications are consistent. It stops if there are any
  errors in the specifications. All of the identified errors are written to
  the run log. A data frame containing the file entries is added to the
  model state file as Geo_df'.
}
\description{
\code{readGeography} a visioneval framework control function that reads the
geographic specifications file for the model.
}
\details{
This function manages the reading and error checking of geographic
specifications for the model. It calls the checkGeography function to check
for errors in the specifications. The checkGeography function reads in the
file and checks for errors. It returns a list of any errors that are found
and a data frame containing the geographic specifications. If errors are
found, the functions writes the errors to a log file and stops model
execution. If there are no errors, the function adds the geographic in the
geographic specifications file, the errors are written to the log file and
execution stops. If no errors are found, the geographic specifications are
added to the model state file.
}
