% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{readModelState}
\alias{readModelState}
\title{Reads values from model state file.}
\usage{
readModelState(Names_ = "All", FileName = "ModelState.Rda")
}
\arguments{
\item{Names_}{A string vector of the components to extract from the
ModelState_ls list.}

\item{FileName}{A string vector with the full path name of the model state
file.}
}
\value{
A list containing the specified components from the model state file.
}
\description{
\code{readModelState} reads components of the file that saves a copy of the
model state
}
\details{
The model state is stored in a list (ModelState_ls) that is also saved as a
file (ModelState.Rda) whenever the list is updated. This function reads the
contents of the ModelState.Rda file.
}

