% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{readVENameRegistry}
\alias{readVENameRegistry}
\title{Reads the VisionEval name registry.}
\usage{
readVENameRegistry(NameRegistryDir = NULL)
}
\arguments{
\item{NameRegistryDir}{a string identifying the path to the directory
where the name registry file is located.}
}
\value{
A list having two components: Inp and Set. Each component is a data
frame containing the respective Inp and Set specifications of registered
modules.
}
\description{
\code{readVENameRegistry} reads the VisionEval name registry and returns a
list of data frames containing the Inp and Set specifications.
}
\details{
The VisionEval name registry (VENameRegistry.json) keeps track of the
dataset names created by all registered modules by reading in datasets
specified in the module Inp specifications or by returning calculated
datasets as specified in the module Set specifications. This function reads
the VisionEval name registry and returns a list of data frames containing the
registered Inp and Set specifications.
}
