% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visioneval.R
\name{runModule}
\alias{runModule}
\title{Run module.}
\usage{
runModule(ModuleName, PackageName, Year, IgnoreInp_ = NULL,
  IgnoreSet_ = NULL)
}
\arguments{
\item{ModuleName}{A string identifying the name of a module object.}

\item{PackageName}{A string identifying the name of the package the module is
a part of.}

\item{Year}{A string identifying the run year.}

\item{IgnoreInp_}{A string vector identifying the names of any inputs
specified by the module that are to be ignored.}

\item{IgnoreSet_}{A string vector identifying the names of any data items in
the module's 'Set' specifications that are to be ignored.}
}
\value{
None. The function writes results to the specified locations in the
  datastore.
}
\description{
\code{runModule} runs a model module.
}
\details{
This function runs a module for  a specified year.
}

