% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{setInDatastore}
\alias{setInDatastore}
\title{Save the data sets returned by a module in the datastore}
\usage{
setInDatastore(Data_ls, ModuleSpec_ls, ModuleName, Geo = NULL)
}
\arguments{
\item{Data_ls}{a list containing the data to be saved. The list is organized
by group, table, and data set.}

\item{ModuleSpec_ls}{a list of module specifications that is consistent with
the VisionEval requirements}

\item{ModuleName}{a string identifying the name of the module (used to document
the module creating the data in the datastore)}

\item{Geo}{a string identifying the name of the geographic area to get the
data for. For example, if the module is specified to be run by Azone, then
Geo would be the name of a particular Azone.}
}
\value{
A list containing all the data sets specified in the module's
'Get' specifications for the identified geographic area.
}
\description{
\code{setInDatastore} saves to the datastore the data returned in a standard
list by a module.
}
\details{
This function saves to the datastore the data sets identified in a module's
'Set' specifications and included in the list returned by the module. If a
particular geographic area is identified, the data are saved to the positions
in the data sets in the datastore corresponding to the identified geographic
area.
}
