% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{sortGeoTable}
\alias{sortGeoTable}
\title{Sort a data frame so that the order of rows matches the geography in a
datastore table.}
\usage{
sortGeoTable(Data_df, Table, Group)
}
\arguments{
\item{Data_df}{a data frame that contains a 'Geo' field containing the names
of the geographic areas to sort by and any number of additional data fields.}

\item{Table}{a string for the table that is to be matched against.}

\item{Group}{a string for the generic group that the table resides in.}
}
\value{
The data frame which has been sorted to match the order of geography
in the specified table in the datastore.
}
\description{
\code{sortGeoTable} returns a data frame whose rows are sorted to match the
geography in a specified table in the datastore.
}
\details{
This function sorts the rows of a data frame that the 'Geo' field in the
data frame matches the corresponding geography names in the specified table
in the datastore. The function returns the sorted table.
}
