% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{Types}
\alias{Types}
\title{Returns a list of returns a list of recognized data types, the units for each
type, and storage mode of each type.}
\usage{
Types()
}
\value{
A list containing a component for each recognized type. Each
component lists the recognized units for the type and the storage mode. There
are currently 4 simple types and 10 complex type. The simple types are
integer, double, character and logical. The complex types are currency,
distance, area, mass, volume, time, speed, vehicle_distance,
passenger_distance, and payload_distance.
}
\description{
\code{Types} returns a list of returns a list of recognized data types, the
units for each type, and storage mode of each type.
}
\details{
This function stores a listing of the dataset types recognized by the
visioneval framework, the units recognized for each type, and the storage
mode used for each type. Types include simple types (e.g. integer, double,
character, logical) as well as complex types (e.g. distance, time, mass). For
the complex types, units are specified as well. For example for the distance
type, allowed units are MI (miles), FT (feet), KM (kilometers), M (meters).
The listing includes conversion factors between units of each complex type.
The listing also contains the storage mode (i.e. integer, double, character,
logical of each type. For simple types, the type and the storage mode are the
same).
}
