% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{writeLog}
\alias{writeLog}
\title{Write to log.}
\usage{
writeLog(Msg = "")
}
\arguments{
\item{Msg}{A character string.}
}
\value{
TRUE if the message is written to the log uccessfully.
It appends the time and the message text to the run log.
}
\description{
\code{writeLog} writes a message to the run log.
}
\details{
This function writes a message in the form of a string to the run log. It
logs the time as well as the message to the run log.
}

