% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{writeToTableRD}
\alias{writeToTableRD}
\title{Write to table.}
\usage{
writeToTableRD(Data_, Spec_ls, Group, Index = NULL)
}
\arguments{
\item{Data_}{A vector of data to be written.}

\item{Spec_ls}{a list containing the standard module 'Set' specifications
described in the model system design documentation.}

\item{Group}{a string representation of the name of the datastore group the
data is to be written to.}

\item{Index}{A numeric vector identifying the positions the data is to be
written to.}
}
\value{
TRUE if data is sucessfully written.
}
\description{
\code{writeToTableRD} writes data to an RData (RD) type datastore table and
initializes dataset if needed.
}
\details{
This function writes a dataset file to an RData (RD) type datastore table. It
initializes the dataset if the dataset does not exist. Enables data to be
written to specific location indexes in the dataset.
}
