% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{writeVENameRegistry}
\alias{writeVENameRegistry}
\title{Writes module Inp and Set specifications to the VisionEval name registry.}
\usage{
writeVENameRegistry(ModuleName, PackageName, NameRegistryDir = NULL)
}
\arguments{
\item{ModuleName}{a string identifying the module name.}

\item{PackageName}{a string identifying the package name.}

\item{NameRegistryDir}{a string identifying the path to the directory
where the name registry file is located.}
}
\value{
TRUE if successful. Has a side effect of updating the VisionEval
name registry.
}
\description{
\code{writeVENameRegistry} writes module Inp and Set specifications to the
VisionEval name registry.
}
\details{
The VisionEval name registry (VENameRegistry.json) keeps track of the
dataset names created by all registered modules by reading in datasets
specified in the module Inp specifications or by returning calculated
datasets as specified in the module Set specifications. This functions adds
the Inp and Set specifications for a module to the registry. It removes any
existing entries for the module first.
}
