% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkDataset}
\alias{checkDataset}
\title{Check dataset existence}
\usage{
checkDataset(Name, Table, Group, DstoreListing_df)
}
\arguments{
\item{Name}{a string identifying the dataset name.}

\item{Table}{a string identifying the table the dataset is a part of.}

\item{Group}{a string or numeric representation of the group the table is a
part of.}

\item{DstoreListing_df}{a dataframe which lists the contents of the datastore
as contained in the model state file.}
}
\value{
A logical identifying whether the dataset is in the datastore
}
\description{
\code{checkDataset} checks whether a dataset exists and if so returns the
full path name.
}
\details{
This function checks whether a dataset exists. The dataset is identified by
its name and the table and group names it is in. If the dataset is not in the
datastore, an error is thrown. If it is located in the datastore, the full
path name to the dataset is returned.
}

