% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{checkModuleSpecs}
\alias{checkModuleSpecs}
\title{Checks module specifications for completeness and for incorrect entries}
\usage{
checkModuleSpecs(Specs_ls, ModuleName)
}
\arguments{
\item{ModuleName}{a string that identifies the name of the module (used for
to identify the module in any error messages.)}

\item{Spec_ls}{a list containing the module specifications as produced
by the 'processModuleSpecs' function.}
}
\value{
A vector containing messages identifying any errors that are found.
}
\description{
\code{checkModuleSpecs}Function checks module specifications for completeness
and for proper values.
}
\details{
This function reads a processed module specifications list and checks whether
the specifications are complete and whether the values are appropriate.
}

