% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{getModelState}
\alias{getModelState}
\title{Retrieve model state.}
\usage{
getModelState(Names_ = "All", FileName = "ModelState.Rda")
}
\arguments{
\item{Names_}{A string vector of the components to extract from the
ModelState_ls list.}

\item{FileName}{A string identifying the name of the file that contains
the ModelState_ls list. The default name is 'ModelState.Rda'.}
}
\value{
A list containing the specified components from the model state file.
}
\description{
\code{getModelState} reads components of the file that keeps track of the
model state
}
\details{
Key variables that are important for managing the model run are stored in a
list (ModelState_ls) that is saved in the 'ModelState.Rda' file. This
function loads the file and extracts named components of the list.
}

