% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{getModelState}
\alias{getModelState}
\title{Get values from model state list.}
\usage{
getModelState(Names_ = "All", State_ls = ModelState_ls)
}
\arguments{
\item{Names_}{A string vector of the components to extract from the
ModelState_ls list.}

\item{State_ls}{The model state list which by default is ModelState_ls.}
}
\value{
A list containing the specified components from the model state file.
}
\description{
\code{getModelState} reads components of the list that keeps track of the
model state
}
\details{
Key variables that are important for managing the model run are stored in a
list (ModelState_ls) that is managed in the global environment. This
function extracts named components of the list.
}
