% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{getUnits}
\alias{getUnits}
\title{Retrieve default units for model}
\usage{
getUnits(Type_)
}
\arguments{
\item{Type_}{A string vector identifying the complex data type(s).}
}
\value{
A string vector identifying the default units for the complex data
type(s) or NA if any of the type(s) are not defined.
}
\description{
\code{getUnits} a visioneval framework control function that retrieves the
default model units for a vector of complex data types.
}
\details{
This is a convenience function to make it easier to retrieve the default
units for a complex data type (e.g. distance, volume, speed). The default
units are the units used to store the complex data type in the datastore.
}
