% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{getYears}
\alias{getYears}
\title{Retrieve years}
\usage{
getYears()
}
\value{
A character vector of the model run years.
}
\description{
\code{getYears} reads the Years component from the the model state file.
}
\details{
This is a convenience function to make it easier to retrieve the Years
component of the model state file.
}

