% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initDataset}
\alias{initDataset}
\title{Initialize dataset in datastore table.}
\usage{
initDataset(Spec_ls, Group)
}
\arguments{
\item{Spec_ls}{a list containing the standard module specifications
described in the model system design documentation.}

\item{Group}{a string representation of the name of the top-level
subdirectory the table is to be created in (i.e. either 'Global' or the name
of the year).}
}
\value{
TRUE if dataset is successfully initialized. If the identified table
does not exist, the function throws an error.
}
\description{
\code{initDataset} initializes a dataset in a table.
}
\details{
This function initializes a dataset which must be done before data can be
stored. The function writes an empty vector to the table directory specified
in the 'TABLE' element of the 'Spec_ls' argument. The type of the vector is
determined by the 'TYPE' element of the 'Spec_ls' argument. The vector is
saved to the table directory with the prefix of the file name being the
'NAME' property of the 'Spec_ls' argument and the suffix being '.Rda'.
}
