% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initDatastore}
\alias{initDatastore}
\title{Initialize Datastore.}
\usage{
initDatastore()
}
\value{
TRUE if datastore initialization is successful. Calls the
listDatastore function which adds a listing of the datastore contents to the
model state file.
}
\description{
\code{initDatastore} creates datastore with starting structure.
}
\details{
This function creates the datastore for the model run with the initial
structure. The datastore is a nested set of directories and files. The
top directory that is named according to the 'DatastoreName' property in the
'run_parameters.json' file. The function creates this directory and several
subdirectories including one named 'Global' and one for each year identified
in the 'Years' property of the 'run_parameters.json' file. If the datastore
directory already exists, the function will remove it and all of its contents
before recreating it.
}
