% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initDatastoreGeography}
\alias{initDatastoreGeography}
\title{Initialize datastore geography.}
\usage{
initDatastoreGeography()
}
\value{
The function returns TRUE if the geographic tables and datasets are
  sucessfully written to the datastore.
}
\description{
\code{initDatastoreGeography} initializes tables and writes datasets to the
datastore which describe geographic relationships of the model.
}
\details{
This function writes tables to the datastore for each of the geographic
levels. These tables are then used during a model run to store values that
are either specified in scenario inputs or that are calculated during a model
run. The function populates the tables with cross-references between
geographic levels. The function reads the model geography (Geo_df) from the
model state file. Upon successful completion, the function calls the
listDatastore function to update the datastore listing in the global list.
}

