% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initDatastoreH5}
\alias{initDatastoreH5}
\title{Initialize Datastore for an HDF5 (H5) type datastore.}
\usage{
initDatastoreH5()
}
\value{
TRUE if datastore initialization is successful. Calls the
listDatastore function which adds a listing of the datastore contents to the
model state file.
}
\description{
\code{initDatastoreH5} creates datastore with starting structure for an HDF5
(H5) type datastore.
}
\details{
This function creates the datastore for the model run with the initial
structure for an HDF5 (H5) type datastore.
}
