% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{initDatastoreRD}
\alias{initDatastoreRD}
\title{Initialize Datastore for an RData (RD) type datastore.}
\usage{
initDatastoreRD()
}
\value{
TRUE if datastore initialization is successful. Calls the
listDatastore function which adds a listing of the datastore contents to the
model state file.
}
\description{
\code{initDatastoreRD} creates datastore with starting structure for an RData
(RD) type datastore.
}
\details{
This function creates the datastore for the model run with the initial
structure for an RData (RD) type datastore.
}
