% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initLog}
\alias{initLog}
\title{Initialize run log.}
\usage{
initLog()
}
\value{
TRUE if the log is created successfully. It creates a log file in the
  working directory and identifies the name of the log file in the
  model state file.
}
\description{
\code{initLog} creates a log (text file) that stores messages generated
during a model run.
}
\details{
This function creates a log file that is a text file which stores messages
generated during a model run. The name of the log is 'Log <date> <time>'
where '<date>' is the initialization date and '<time>' is the initialization
time. The log is initialized with the scenario name, scenario description and
the date and time of initialization.
}

