% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initModelStateFile}
\alias{initModelStateFile}
\title{Initialize model state.}
\usage{
initModelStateFile(Dir = "defs", ParamFile = "run_parameters.json")
}
\arguments{
\item{Dir}{A string identifying the name of the directory where the global
parameters file is located. The default value is "defs".}

\item{ParamFile}{A string identifying the name of the global parameters file.
The default value is "parameters.json".}
}
\value{
TRUE if the model state list is created and file is saved. It creates
the model state list and loads parameters recorded in the 'parameters.json'
file into the model state lists and saves a model state file.
}
\description{
\code{initModelState} loads model run parameters into the model state
list in the global workspace and saves as file.
}
\details{
This function creates the model state list and loads model run parameters
recorded in the 'parameters.json' file into the model state list. It also
saves the model state list in a file (ModelState.Rda).
}

