% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{initTable}
\alias{initTable}
\title{Initialize table in datastore.}
\usage{
initTable(Table, Group, Length)
}
\arguments{
\item{Group}{a string representation of the name of the group the table is to
be created in.}

\item{Spec_ls}{a list containing the standard module 'Set' specifications
described in the model system design documentation.}
}
\value{
The value TRUE is returned if the function is successful at creating
  the table. In addition, the listDatastore function is run to update the
  inventory in the model state file. The function stops if the group in which
  the table is to be placed does not exist in the datastore and a message is
  written to the log.
}
\description{
\code{initDatastoreTable} initializes a table in the datastore.
}
\details{
A table in the datastore is a group which contains equal length vectors of
data that may have different types. Thus it is much like an R data frame
which is a list containing equal length vectors that may have different
types. A table is initialized in the datastore by initializing a group whose
name is the table name and assigning a 'LENGTH' attribute which specifies the
length that all data vectors in the table must have.
}

