% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{listDatastore}
\alias{listDatastore}
\title{List datastore contents.}
\usage{
listDatastore()
}
\value{
TRUE if the listing is successfully read from the datastore and
written to the model state file.
}
\description{
\code{listDatastore} lists the contents of a datastore.
}
\details{
This function lists the contents of a datastore including identifying all
groups, tables, and datasets. It also lists the attributes associated with
each table and dataset. The listing is stored in the global list as
'G$Datastore'. This function is run whenever the structure or contents of the
datastore is changed to always keep the listing in G$Datastore current.
}
