% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{parseModelScript}
\alias{parseModelScript}
\title{Parse model script.}
\usage{
parseModelScript(FilePath = "run_model.R", TestMode = FALSE)
}
\arguments{
\item{FilePath}{A string identifying the relative or absolute path to the
model run script is located.}

\item{TestMode}{A logical identifying whether the function is to run in test
mode. When in test mode the function returns the parsed script but does not
change the model state or write results to the log.}
}
\value{
A data frame containing information on the calls to 'runModule' in the
order of the calls. Each row represents a module call in order. The columns
identify the 'ModuleName', the 'PackageName', and the 'RunFor' value.
}
\description{
\code{parseModel} a visioneval framework control function that reads and
parses the model script to identify the sequence of module calls and the
associated call arguments.
}
\details{
This function reads in the model run script and parses the script to
identify the sequence of module calls. It extracts each call to 'runModule'
and identifies the values assigned to the function arguments. It creates a
list of the calls with their arguments in the order of the calls in the
script.
}
