% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{processModuleInputs}
\alias{processModuleInputs}
\title{Write inputs to datastore.}
\usage{
processModuleInputs(Inp_ls, ModuleName, Dir = "inputs", OnlyCheck = TRUE,
  Ignore_ = NULL)
}
\arguments{
\item{Inp_ls}{A list which meets standards for specifying input
characteristics.}

\item{ModuleName}{A string identifying the module name.}

\item{Dir}{A string identifying the relative path name to the directory where
input files are located.}

\item{OnlyCheck}{A logical value. If TRUE, the function will only check the
inputs. If FALSE, the function will check the inputs and load the input
data into the datastore.}

\item{Ignore_}{A vector of the names of data items that will not be checked
or loaded. If NULL, all data items will be checked and loaded.}
}
\value{
A numeric vector having 3 elements (FileErrors, DataErrors,
  DataWarnings). Each element identifies how may data errors or warnings were
  found.
}
\description{
\code{processModuleInputs} checks whether specified module inputs meet
specifications and load into datastore if OnlyCheck argument is false.
}
\details{
This function checks whether all the specified scenario inputs meet
specifications. Checks include whether the specified files exist, whether the
files include data for all years and specified zones, and whether every data
item meets specifications. Data specifications.
}

