% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{readFromTable}
\alias{readFromTable}
\title{Read from table.}
\usage{
readFromTable(Name, Table, Group, File = "datastore.h5", Index = NULL)
}
\arguments{
\item{Name}{A string identifying the name of the dataset to be read from.}

\item{Table}{A string identifying the complete name of the table where the
dataset is located.}

\item{Group}{a string representation of the name of the datastore group the
data is to be read from.}

\item{File}{a string representation of the file path of the datastore}

\item{Index}{A numeric vector identifying the positions the data is to be
written to. NULL if the entire dataset is to be read.}
}
\value{
A vector of the same type stored in the datastore and specified in
  the TYPE attribute.
}
\description{
\code{readFromTable} reads a dataset from a table.
}
\details{
This function reads datasets from a table. Indexed reads are permitted. The
function checks whether the table and dataset exist and whether all specified
indices are within the length of the table. The function converts any values
equal to the NAVALUE attribute to NA.
}
