% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visioneval.R
\name{runModule}
\alias{runModule}
\title{Run module.}
\usage{
runModule(ModuleName, PackageName, RunFor, RunYear = Year)
}
\arguments{
\item{ModuleName}{A string identifying the name of a module object.}

\item{PackageName}{A string identifying the name of the package the module is
a part of.}

\item{RunFor}{A string identifying whether to run the module for all years
"AllYears", only the base year "BaseYear", or for all years except the base
year "NotBaseYear".}

\item{RunYear}{A string identifying the run year.}
}
\value{
None. The function writes results to the specified locations in the
  datastore and prints a message to the console when the module is being run.
}
\description{
\code{runModule} runs a model module.
}
\details{
This function runs a module for  a specified year.
}

